Changed id from Integer to Int, implemented TreeController, added AbstractEntity

+++ family-tree-backend/src/main/java/com/example/tree/repositories/PersonRepository.java
@@ -3,5 +3,5 @@ package com.example.tree.repositories;
 import com.example.tree.model.Person;
 import org.springframework.data.repository.CrudRepository;
 
-public interface PersonRepository extends CrudRepository<Person, Integer> {
+public interface PersonRepository extends CrudRepository<Person, Long> {
 }
